add the missing MIT LICENSE file
there is no LICENSE file in the repository. You do not need to copy the complete license to each header. It may be useful to add a single line in the header. For people who distribute synctex (like developers in Linux distributions) it helps a lot to have the LICENSE file where it is expected to be. I prepared Line 3 for you, please adjust it to your taste. Best, JS