Use clang-format (#98)
* Strip trailing whitespace

* Fix up README

* Fix include directive

* Wrap table-like formatting in `clang-format off`

* Prepare for clang-format: add trailing comma to struct initializers

It prevents clang-format from indenting the fields all the way to the
right.

* Prepare for clang-format: Move defines out of the middle of a function call

* Remove unused define

* Add semicolons to top-level standalone macro calls

While not syntactically required *after* macro expansion, in the form of
source code *before* the expansion it makes syntax highlighters and
clang-format quite confused that those macros might expand to some
attributes for a next function or variable definition, and it messes up
colors/indentation.

* Add clang-format configuration

* cd meson; meson setup --wipe build; ninja -C build clang-format

---------

Co-authored-by: ivan tkachenko <me@ratijas.me>